[MNG-6110] renamed 'maven-aether-provider' to 'maven-resolver-provider'
