+++ maven-resolver-provider/src/main/java/org/apache/maven/repository/internal/MavenMetadata.java
