+++ maven-resolver-provider/src/site/site.xml
