+++ maven-resolver-provider/src/test/resources/repo/ut/simple/artifact/1.0/artifact-1.0.pom
