+++ maven-resolver-provider/src/test/resources/repo/ut/simple/artifact/maven-metadata.xml
