+++ maven-resolver-provider/src/test/resources/repo/ut/simple/dependency/1.0/dependency-1.0.pom
