+++ maven-resolver-provider/src/test/resources/repo/ut/simple/dependency/maven-metadata.xml
