+++ maven-resolver-provider/src/test/resources/repo/ut/simple/parent/1.0/parent-1.0.pom
