+++ maven-resolver-provider/src/test/resources/repo/ut/simple/parent/maven-metadata.xml
